requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
lxml==4.5.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.23.0
requests-cache==0.5.2
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
